# api/swagger.yaml
info:
  contact:
    email: support@openusp.io
    name: OpenUSP Support
    url: https://github.com/plume-design-inc/openusp
  description: OpenUSP API Gateway provides RESTful APIs for managing TR-369 USP and
    TR-069 CWMP devices
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OpenUSP API Gateway
  version: "1.0"
paths:
  /api/v1/alerts:
    get:
      description: Get a list of all alerts in the system
      parameters:
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Pagination limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all alerts
      tags:
      - Alerts
    post:
      consumes:
      - application/json
      description: Create a new alert
      parameters:
      - description: Alert data
        in: body
        name: alert
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create alert
      tags:
      - Alerts
  /api/v1/alerts/{id}/resolve:
    put:
      description: Mark an alert as resolved
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Resolve alert
      tags:
      - Alerts
  /api/v1/devices:
    get:
      description: Get a list of all registered devices
      parameters:
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Pagination limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all devices
      tags:
      - Devices
    post:
      consumes:
      - application/json
      description: Register a new device in the system
      parameters:
      - description: Device information
        in: body
        name: device
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new device
      tags:
      - Devices
  /api/v1/devices/{device_id}:
    delete:
      description: Remove a device from the system
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete device
      tags:
      - Devices
    get:
      description: Get detailed information about a specific device
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get device by ID
      tags:
      - Devices
    put:
      consumes:
      - application/json
      description: Update device information
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Device information
        in: body
        name: device
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update device
      tags:
      - Devices
  /api/v1/devices/{device_id}/alerts:
    get:
      description: Get all alerts for a specific device
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get device alerts
      tags:
      - Alerts
  /api/v1/devices/{device_id}/commands/execute:
    post:
      consumes:
      - application/json
      description: Execute a command on a device
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Command data
        in: body
        name: command
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Execute command
      tags:
      - Commands
  /api/v1/devices/{device_id}/datamodel:
    get:
      description: Retrieve the complete data model for a device
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get device data model
      tags:
      - DataModel
  /api/v1/devices/{device_id}/objects:
    get:
      description: Retrieve all objects for a device
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get device objects
      tags:
      - Objects
  /api/v1/devices/{device_id}/objects/{object_path}:
    get:
      description: Retrieve a specific object by its path
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Object path
        in: path
        name: object_path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get object by path
      tags:
      - Objects
...
  /api/v1/devices/{device_id}/parameters/search:
    get:
      description: Search for device parameters by path
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Parameter path to search
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search parameters
      tags:
      - Parameters
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication. Use "Bearer <token>" format.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"